compiler.h: eliminate some unnecessary abstraction layers
wimlib only supports being built with gcc or clang.  This is not planned
to change, as MSVC support would be a lot of work.  Many of the
compiler-related abstraction layers (in compiler.h) can't simply be
implemented for MSVC; the users will need to be changed too.

Therefore, let's just remove some of these unnecessary abstraction
layers so that it's clearer what is really going on in the code.